Crate drone_macros_core

Procedural macros base for Drone, an Embedded Operating System.

This crate provides shared functionality for all Drone procedural macro crates.

Macros

compile_error
Unconditionally causes parsing to fail with the given error message.

new_ident
Creates call site struct@syn::Ident using interpolation of runtime expressions.

Structs

CfgCond
List of conditional compilation clauses.

Traits

CfgCondExt
CfgCond helper extension trait for slices.

Functions

unkeywordize
Inserts an underscore at the beginning of the string if the string is a reserved keyword.